Extract String Resource For All Text Strings And Replace All Text Strings With String Resources
Certain text strings in the app were not extracted as string resource, which was messing with the translation process as certain text strings will not show up in the Translation Editor.

For all text strings, I extracted string resource if the equivalent text was not extracted previously. For text strings that have existing equivalent string resource, I replace the text strings with the string resource.